make it use langdb.js
I don't know what the best way to do this is. I could generate
just the data needed for each article and insert it in that
article. Instead I put all the data for all articles in /langdb.js

Then at runtime I generate a language selector.

When the user selects a language, if the current article
exists for that language I go to that article. If it doesn't
exist I go the index page for that language since presumably
the index page lists what articles are available in that language