fix issue with debug wrapper
The original wrapper created a new object instead of modifying the
existing one. This turned out to have consequences.

In particular the user has a reference to the wrapper. So we couldn't
actually restore the function to their original functions, instead we
had to patch the wrapper to have smaller wrapper functions (functions
that didn't check for errors). The code to do that had the issue that
for every function found on the wrapper it would create a function
that called the same named function on the original. But, if someone
added a function to the wrapper (meaning they were adding functions
to the context) then, when "unwrapping" there would be an error where

   wrapper[nameOfFunc] = origContext[nameOfFunc].bind(origContext)

Because nameOfFunc doesn't exist on the original context.