Fix multiple-views-items examples
I have no idea when this broke or why it broke. If you run the
examples outside the editor they work. If you export them to
jsfiddle, codepen, jsgist, stackoverflow, they all work.
But for some reason they don't work in these iframes.

To see the issue, reset to the previous version of this
repo, build, view the webgl/lessons/webgl-multiple-views.html
lesson. Scroll down to the last two examples. In the example
itself, scroll down. Eventually the 3D stuff will disappear.

The issue is the browser stops updating window.scrollY at
some point. I don't know why. As it says above, outside an
iframe and even inside an iframe on all those sites it works
fine.

Here it's an iframe in an iframe but I'm not convinced that's
it as I think stackoverflow's snippet editor is an iframe in
an iframe.

I was able to repo at least one issue in an iframe
but it didn't lead to what the problem is.

Using an outer div and scrolling inside that seems to fix
the issue.