Revert "Fix multiple-views-items examples"
This reverts commit 9651f2539e6c5c795873600f8d086673c6d5bc9e.

The issue was not the CSS. It was the code trying to fix
requestAnimationFrame so that it only fires if the iframe
is visible. The code worked when it shipped but browsers
broke it.

Fortunately browsers no longer run requestAnimationFrame
in an iframe if the iframe is not visible so the code
is no longer needed.